backend/requirements.txt
pyramid==2.0.2
waitress==3.0.0
pyramid_tm==2.5
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
transformers==4.44.2
torch>=2.1.0
google-generativeai==0.7.2
